@@ -134,119 +66,124 @@
     ]
   },
   "devDependencies": {
-    "@heroicons/react": "^1.0.5",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.5.1",
-    "@svgr/core": "^6.2.0",
-    "@teamsupercell/typings-for-css-modules-loader": "^2.5.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@types/history": "4.7.9",
-    "@types/jest": "^27.0.2",
-    "@types/lodash": "^4.14.184",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-router": "^5.1.17",
-    "@types/react-router-dom": "^5.3.2",
-    "@types/react-test-renderer": "^17.0.1",
-    "@types/terser-webpack-plugin": "^5.0.4",
-    "@types/webpack-env": "^1.16.3",
-    "@typescript-eslint/eslint-plugin": "^5.2.0",
-    "autoprefixer": "^10.4.0",
-    "browserslist-config-erb": "^0.0.3",
-    "chalk": "^4.1.2",
-    "concurrently": "^6.3.0",
-    "core-js": "^3.19.0",
+    "@babel/cli": "^7.28.0",
+    "@babel/core": "^7.28.0",
+    "@babel/preset-env": "^7.28.0",
+    "@babel/preset-react": "^7.27.1",
+    "@babel/preset-typescript": "^7.27.1",
+    "@electron-forge/cli": "^7.8.1",
+    "@electron-forge/maker-deb": "^7.8.1",
+    "@electron-forge/maker-rpm": "^7.8.1",
+    "@electron-forge/maker-squirrel": "^7.8.1",
+    "@electron-forge/maker-zip": "^7.8.1",
+    "@electron-forge/plugin-auto-unpack-natives": "^7.8.1",
+    "@electron-forge/plugin-fuses": "^7.8.1",
+    "@electron-forge/plugin-webpack": "^7.8.1",
+    "@electron/fuses": "^1.8.0",
+    "@heroicons/react": "^2.2.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "0.6.1",
+    "@svgr/core": "^8.1.0",
+    "@teamsupercell/typings-for-css-modules-loader": "^2.5.2",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@types/history": "5.0.0",
+    "@types/jest": "^30.0.0",
+    "@types/lodash": "^4.17.20",
+    "@types/node": "^24.0.10",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-test-renderer": "^19.1.0",
+    "@types/terser-webpack-plugin": "^5.2.0",
+    "@types/webpack-bundle-analyzer": "^4.7.0",
+    "@types/webpack-env": "^1.18.8",
+    "@typescript-eslint/eslint-plugin": "^8.35.1",
+    "autoprefixer": "^10.4.21",
+    "babel-loader": "^10.0.0",
+    "chalk": "^5.4.1",
+    "concurrently": "^9.2.0",
+    "core-js": "^3.43.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.5.0",
-    "css-minimizer-webpack-plugin": "^3.1.1",
-    "detect-port": "^1.3.0",
-    "dotenv": "^16.0.0",
-    "electron": "^15.3.0",
-    "electron-builder": "^23.0.2",
-    "electron-devtools-installer": "^3.2.0",
-    "electron-notarize": "^1.1.1",
-    "electron-rebuild": "^3.2.3",
-    "electron-winstaller": "^5.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-erb": "^4.0.3",
-    "eslint-plugin-compat": "^3.13.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jest": "^25.2.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "css-loader": "^7.1.2",
+    "css-minimizer-webpack-plugin": "^7.0.2",
+    "detect-port": "^2.1.0",
+    "dotenv": "^17.0.1",
+    "electron": "^37.2.0",
+    "electron-devtools-installer": "^4.0.0",
+    "electron-notarize": "^1.2.2",
+    "electron-rebuild": "^3.2.9",
+    "eslint": "^9.30.1",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "7.0.4",
+    "html-webpack-plugin": "^5.6.3",
+    "husky": "9.1.7",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.3.1",
-    "lint-staged": "^11.2.6",
-    "mini-css-extract-plugin": "^2.4.3",
+    "jest": "^30.0.4",
+    "lint-staged": "^16.1.2",
+    "mini-css-extract-plugin": "^2.9.2",
     "opencollective-postinstall": "^2.0.3",
-    "postcss": "^8.3.11",
-    "postcss-loader": "^6.2.0",
-    "prettier": "^2.4.1",
-    "react-refresh": "^0.10.0",
-    "react-refresh-typescript": "^2.0.2",
-    "react-test-renderer": "^17.0.2",
-    "rimraf": "^3.0.2",
-    "sass": "^1.43.4",
-    "sass-loader": "^12.3.0",
-    "style-loader": "^3.3.1",
-    "tailwindcss": "^2.2.19",
-    "terser-webpack-plugin": "^5.2.4",
-    "ts-jest": "^27.0.7",
-    "ts-loader": "^9.2.6",
-    "ts-node": "10.8.1",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^4.4.4",
+    "postcss": "^8.5.6",
+    "postcss-loader": "^8.1.1",
+    "prettier": "^3.6.2",
+    "react-refresh": "^0.17.0",
+    "react-refresh-typescript": "^2.0.10",
+    "react-test-renderer": "^19.1.0",
+    "rimraf": "^6.0.1",
+    "sass": "^1.89.2",
+    "sass-loader": "^16.0.5",
+    "style-loader": "^4.0.0",
+    "tailwindcss": "^4.1.11",
+    "terser-webpack-plugin": "^5.3.14",
+    "ts-jest": "^29.4.0",
+    "ts-node": "^10.9.2",
+    "tsx": "^4.20.3",
+    "typescript": "^5.8.3",
     "url-loader": "^4.1.1",
-    "webpack": "^5.60.0",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.8.0"
+    "webpack-bundle-analyzer": "^4.10.2",
+    "webpack-dev-server": "^5.2.2",
+    "webpack-merge": "^6.0.1"
   },
   "dependencies": {
-    "@electron/remote": "^2.0.4",
-    "@headlessui/react": "^1.6.6",
+    "@electron/remote": "^2.1.2",
+    "@headlessui/react": "^2.2.4",
     "@node-steam/vdf": "^2.2.0",
     "@redux-beacon/google-analytics": "^1.2.3",
-    "@reduxjs/toolkit": "^1.7.1",
-    "axios": "^0.24.0",
+    "@reduxjs/toolkit": "^2.8.2",
+    "@vercel/webpack-asset-relocator-loader": "^1.7.3",
+    "axios": "^1.10.0",
     "bytebuffer": "^5.0.1",
-    "chart.js": "^3.8.0",
-    "chartjs-plugin-datalabels": "^2.0.0",
-    "currency-converter-lt": "^1.2.4",
-    "electron-debug": "^3.2.0",
-    "electron-dl": "^3.3.0",
-    "electron-log": "^4.4.8",
-    "electron-redux": "^1.5.4",
-    "electron-store": "^8.0.1",
-    "electron-updater": "^4.3.9",
-    "fernet": "^0.4.0",
-    "find-process": "^1.4.7",
+    "chart.js": "^4.5.0",
+    "chartjs-plugin-datalabels": "^2.2.0",
+    "currency-converter-lt": "^1.3.1",
+    "electron-debug": "^4.1.0",
+    "electron-dl": "^4.0.0",
+    "electron-log": "^5.4.1",
+    "electron-redux": "^2.0.0",
+    "electron-squirrel-startup": "^1.0.1",
+    "electron-store": "^10.1.0",
+    "electron-updater": "^6.6.2",
+    "fernet": "^0.3.3",
+    "find-process": "^2.0.0",
+    "fork-ts-checker-webpack-plugin": "^9.1.0",
     "globaloffensive": "^3.2.0",
-    "history": "4.x.x",
+    "history": "5.x.x",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-chartjs-2": "^4.2.0",
-    "react-dom": "^17.0.2",
-    "react-qr-code": "^2.0.11",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
+    "react": "^19.1.0",
+    "react-chartjs-2": "^5.3.0",
+    "react-dom": "^19.1.0",
+    "react-qr-code": "^2.0.17",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^7.6.3",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.4.1",
-    "regenerator-runtime": "^0.13.9",
-    "steam-session": "^1.1.1",
-    "steam-totp": "^2.1.1",
-    "steam-user": "^4.28.1",
-    "update-electron-app": "^2.0.1",
-    "web-streams-polyfill": "^4.1.0",
+    "redux-thunk": "^3.1.0",
+    "regenerator-runtime": "^0.14.1",
+    "steam-session": "^1.9.3",
+    "steam-totp": "^2.1.2",
+    "steam-user": "^5.2.3",
+    "uglify-js": "^3.19.3",
+    "update-electron-app": "^3.1.1",
+    "webpack": "^5.100.1",
+    "webpack-cli": "^6.0.1",
     "xmldom": "^0.6.0"
   },
   "overrides": {
