@@ -1,10 +1,10 @@
-import { render } from "react-dom";
+import { createRoot } from 'react-dom/client';
 import { Provider } from "react-redux";
 import App from "./App";
 import { PersistGate } from 'redux-persist/integration/react'
 import returnVar from './store/configureStore'
 
-const myVar = returnVar()
+const myVar = returnVar();
 
 
 declare global {
@@ -21,11 +21,15 @@ declare global {
   }
 }
 
-render(
 
-  <Provider store={myVar.reduxStore}>
+const container = document.getElementById('root');
+if (container != null) {
+const root = createRoot(container);
+root.render(
+   <Provider store={myVar.reduxStore}>
     <PersistGate loading={null} persistor={myVar.persistor}>
      <App />
      </PersistGate>
   </Provider>
-, document.getElementById("root"));
+);
+}
